Mission: Split out classes to their own files
The Mission file was getting a bit large, so make it its own module and
move classes to their own files. This means that during missions we only
load the mission (super)classes that we need.
