+++ mission/__init__.py
@@ -1,0 +1,6 @@
+# We only list usable missions here, not base classes.
+__all__ = [
+    "Mission_Browse", "Mission_Calibrate", "Mission_Cycle", "Mission_Forward",
+    "Mission_Infrared", "Mission_Infrared_Grid", "Mission_Pathfind",
+    "Mission_Search", "Mission_Square", "Mission_XBee"
+]
